rename plotting --> chartly
