+++ docs/py-modindex.html
@@ -68,24 +68,24 @@
    <h1>Python Module Index</h1>
 
    <div class="modindex-jumpbox">
-   <a href="#cap-p"><strong>p</strong></a> | 
+   <a href="#cap-c"><strong>c</strong></a> | 
    <a href="#cap-t"><strong>t</strong></a>
    </div>
 
    <table class="indextable modindextable">
      <tr class="pcap"><td></td><td>&#160;</td><td></td></tr>
-     <tr class="cap" id="cap-p"><td></td><td>
-       <strong>p</strong></td><td></td></tr>
+     <tr class="cap" id="cap-c"><td></td><td>
+       <strong>c</strong></td><td></td></tr>
      <tr>
        <td><img src="_static/minus.png" class="toggler"
               id="toggle-1" style="display: none" alt="-" /></td>
        <td>
-       <a href="autoapi/plotting/index.html#module-plotting"><code class="xref">plotting</code></a></td><td>
+       <a href="autoapi/chartly/index.html#module-chartly"><code class="xref">chartly</code></a></td><td>
        <em></em></td></tr>
      <tr class="cg-1">
        <td></td>
        <td>&#160;&#160;&#160;
-       <a href="autoapi/plotting/plotting/index.html#module-plotting.plotting"><code class="xref">plotting.plotting</code></a></td><td>
+       <a href="autoapi/chartly/chartly/index.html#module-chartly.chartly"><code class="xref">chartly.chartly</code></a></td><td>
        <em></em></td></tr>
      <tr class="pcap"><td></td><td>&#160;</td><td></td></tr>
      <tr class="cap" id="cap-t"><td></td><td>
@@ -99,7 +99,7 @@
      <tr class="cg-2">
        <td></td>
        <td>&#160;&#160;&#160;
-       <a href="autoapi/tests/test_plotting/index.html#module-tests.test_plotting"><code class="xref">tests.test_plotting</code></a></td><td>
+       <a href="autoapi/tests/test_chartly/index.html#module-tests.test_chartly"><code class="xref">tests.test_chartly</code></a></td><td>
        <em></em></td></tr>
    </table>
