+++ docs/searchindex.js
@@ -1,0 +1,1 @@
+Search.setIndex({"alltitles": {"API Reference": [[4, null]], "Attributes": [[2, "attributes"]], "Chartly": [[7, null]], "Chartly.Multiplots Usage Guide": [[0, null]], "Classes": [[2, "classes"], [6, "classes"]], "Contour Plot": [[1, "contour-plot"]], "Examples": [[0, "examples"], [1, "examples"]], "Generic Plot": [[1, "generic-plot"]], "Histogram": [[1, "histogram"]], "Indices and tables": [[7, "indices-and-tables"]], "Installation": [[7, "installation"]], "Module Contents": [[2, "module-contents"], [6, "module-contents"]], "Normal Cumulative Distribution Function Plot": [[1, "normal-cumulative-distribution-function-plot"]], "Normal Probability Plot": [[1, "normal-probability-plot"]], "Overlay Plots": [[0, "overlay-plots"]], "Overview": [[7, "overview"]], "Requirements": [[7, "requirements"]], "Submodules": [[3, "submodules"], [5, "submodules"]], "Subplots": [[0, "subplots"]], "Usage": [[7, "usage"]], "chartly": [[3, null]], "chartly.Plot Usage Guide": [[1, null]], "chartly.chartly": [[2, null]], "tests": [[5, null]], "tests.test_chartly": [[6, null]]}, "docnames": ["Multiplots", "Plot", "autoapi/chartly/chartly/index", "autoapi/chartly/index", "autoapi/index", "autoapi/tests/index", "autoapi/tests/test_chartly/index", "index"], "envversion": {"sphinx": 64, "sphinx.domains.c": 3, "sphinx.domains.changeset": 1, "sphinx.domains.citation": 1, "sphinx.domains.cpp": 9, "sphinx.domains.index": 1, "sphinx.domains.javascript": 3, "sphinx.domains.math": 2, "sphinx.domains.python": 4, "sphinx.domains.rst": 2, "sphinx.domains.std": 2}, "filenames": ["Multiplots.rst", "Plot.rst", "autoapi/chartly/chartly/index.rst", "autoapi/chartly/index.rst", "autoapi/index.rst", "autoapi/tests/index.rst", "autoapi/tests/test_chartly/index.rst", "index.rst"], "indexentries": {"__call__() (chartly.chartly.multiplots method)": [[2, "chartly.chartly.Multiplots.__call__", false]], "ax (chartly.chartly.plot attribute)": [[2, "chartly.chartly.Plot.ax", false]], "axes_labels (chartly.chartly.plot attribute)": [[2, "chartly.chartly.Plot.axes_labels", false]], "boxplot_args (chartly.chartly.plot attribute)": [[2, "chartly.chartly.Plot.boxplot_args", false]], "cdf_args (chartly.chartly.plot attribute)": [[2, "chartly.chartly.Plot.cdf_args", false]], "chartly": [[3, "module-chartly", false]], "chartly.chartly": [[2, "module-chartly.chartly", false]], "clear_axis() (chartly.chartly.multiplots method)": [[2, "chartly.chartly.Multiplots.clear_axis", false]], "color_list (in module chartly.chartly)": [[2, "chartly.chartly.color_list", false]], "contour_args (chartly.chartly.plot attribute)": [[2, "chartly.chartly.Plot.contour_args", false]], "current_subplot (chartly.chartly.multiplots attribute)": [[2, "chartly.chartly.Multiplots.current_subplot", false]], "data (chartly.chartly.plot attribute)": [[2, "chartly.chartly.Plot.data", false]], "density_args (chartly.chartly.plot attribute)": [[2, "chartly.chartly.Plot.density_args", false]], "export_args (chartly.chartly.plot attribute)": [[2, "chartly.chartly.Plot.export_args", false]], "export_plot_to_file() (chartly.chartly.plot method)": [[2, "chartly.chartly.Plot.export_plot_to_file", false]], "fig (chartly.chartly.plot attribute)": [[2, "chartly.chartly.Plot.fig", false]], "gen_plot_args (chartly.chartly.plot attribute)": [[2, "chartly.chartly.Plot.gen_plot_args", false]], "generic_plot() (chartly.chartly.plot method)": [[2, "chartly.chartly.Plot.generic_plot", false]], "graphs (chartly.chartly.multiplots attribute)": [[2, "chartly.chartly.Multiplots.graphs", false]], "hist_args (chartly.chartly.plot attribute)": [[2, "chartly.chartly.Plot.hist_args", false]], "label_axes() (chartly.chartly.plot method)": [[2, "chartly.chartly.Plot.label_axes", false]], "module": [[2, "module-chartly.chartly", false], [3, "module-chartly", false], [5, "module-tests", false], [6, "module-tests.test_chartly", false]], "multiplots (class in chartly.chartly)": [[2, "chartly.chartly.Multiplots", false]], "new_subplot() (chartly.chartly.multiplots method)": [[2, "chartly.chartly.Multiplots.new_subplot", false]], "norm_cdf_args (chartly.chartly.plot attribute)": [[2, "chartly.chartly.Plot.norm_cdf_args", false]], "overlay() (chartly.chartly.multiplots method)": [[2, "chartly.chartly.Multiplots.overlay", false]], "plot (class in chartly.chartly)": [[2, "chartly.chartly.Plot", false]], "plot_box_plots() (chartly.chartly.plot method)": [[2, "chartly.chartly.Plot.plot_box_plots", false]], "plot_cdf() (chartly.chartly.plot method)": [[2, "chartly.chartly.Plot.plot_cdf", false]], "plot_contour() (chartly.chartly.plot method)": [[2, "chartly.chartly.Plot.plot_contour", false]], "plot_density() (chartly.chartly.plot method)": [[2, "chartly.chartly.Plot.plot_density", false]], "plot_histogram() (chartly.chartly.plot method)": [[2, "chartly.chartly.Plot.plot_histogram", false]], "plot_normal_cdf() (chartly.chartly.plot method)": [[2, "chartly.chartly.Plot.plot_normal_cdf", false]], "plot_probability_plot() (chartly.chartly.plot method)": [[2, "chartly.chartly.Plot.plot_probability_plot", false]], "plot_timeseries() (chartly.chartly.plot method)": [[2, "chartly.chartly.Plot.plot_timeseries", false]], "plotutilities (class in chartly.chartly)": [[2, "chartly.chartly.PlotUtilities", false]], "prob_plot_args (chartly.chartly.plot attribute)": [[2, "chartly.chartly.Plot.prob_plot_args", false]], "set_defaults() (chartly.chartly.plot method)": [[2, "chartly.chartly.Plot.set_defaults", false]], "setup() (tests.test_chartly.testplotting method)": [[6, "tests.test_chartly.TestPlotting.setUp", false]], "share_axes (chartly.chartly.multiplots attribute)": [[2, "chartly.chartly.Multiplots.share_axes", false]], "standardize_dataset() (chartly.chartly.plotutilities method)": [[2, "chartly.chartly.PlotUtilities.standardize_dataset", false]], "subplot_count (chartly.chartly.multiplots attribute)": [[2, "chartly.chartly.Multiplots.subplot_count", false]], "subplots (chartly.chartly.multiplots attribute)": [[2, "chartly.chartly.Multiplots.subplots", false]], "super_title (chartly.chartly.multiplots attribute)": [[2, "chartly.chartly.Multiplots.super_title", false]], "super_xlabel (chartly.chartly.multiplots attribute)": [[2, "chartly.chartly.Multiplots.super_xlabel", false]], "super_ylabel (chartly.chartly.multiplots attribute)": [[2, "chartly.chartly.Multiplots.super_ylabel", false]], "test_contour_data_length() (tests.test_chartly.testplotting method)": [[6, "tests.test_chartly.TestPlotting.test_contour_data_length", false]], "test_default() (tests.test_chartly.testplotting method)": [[6, "tests.test_chartly.TestPlotting.test_default", false]], "test_gen_plot_data_length() (tests.test_chartly.testplotting method)": [[6, "tests.test_chartly.TestPlotting.test_gen_plot_data_length", false]], "test_gen_plot_data_type() (tests.test_chartly.testplotting method)": [[6, "tests.test_chartly.TestPlotting.test_gen_plot_data_type", false]], "test_multi_clear_axis() (tests.test_chartly.testplotting method)": [[6, "tests.test_chartly.TestPlotting.test_multi_clear_axis", false]], "test_mutli_subplot_count() (tests.test_chartly.testplotting method)": [[6, "tests.test_chartly.TestPlotting.test_mutli_subplot_count", false]], "test_standardize_data() (tests.test_chartly.testplotting method)": [[6, "tests.test_chartly.TestPlotting.test_standardize_data", false]], "testplotting (class in tests.test_chartly)": [[6, "tests.test_chartly.TestPlotting", false]], "tests": [[5, "module-tests", false]], "tests.test_chartly": [[6, "module-tests.test_chartly", false]], "tiling() (chartly.chartly.multiplots method)": [[2, "chartly.chartly.Multiplots.tiling", false]], "timeseries_args (chartly.chartly.plot attribute)": [[2, "chartly.chartly.Plot.timeseries_args", false]], "update_defaults() (chartly.chartly.plot method)": [[2, "chartly.chartly.Plot.update_defaults", false]], "util (chartly.chartly.plot attribute)": [[2, "chartly.chartly.Plot.util", false]]}, "objects": {"": [[3, 0, 0, "-", "chartly"], [5, 0, 0, "-", "tests"]], "chartly": [[2, 0, 0, "-", "chartly"]], "chartly.chartly": [[2, 1, 1, "", "Multiplots"], [2, 1, 1, "", "Plot"], [2, 1, 1, "", "PlotUtilities"], [2, 4, 1, "", "color_list"]], "chartly.chartly.Multiplots": [[2, 2, 1, "", "__call__"], [2, 2, 1, "", "clear_axis"], [2, 3, 1, "", "current_subplot"], [2, 3, 1, "", "graphs"], [2, 2, 1, "", "new_subplot"], [2, 2, 1, "", "overlay"], [2, 3, 1, "", "share_axes"], [2, 3, 1, "", "subplot_count"], [2, 3, 1, "", "subplots"], [2, 3, 1, "", "super_title"], [2, 3, 1, "", "super_xlabel"], [2, 3, 1, "", "super_ylabel"], [2, 2, 1, "", "tiling"]], "chartly.chartly.Plot": [[2, 3, 1, "", "ax"], [2, 3, 1, "", "axes_labels"], [2, 3, 1, "", "boxplot_args"], [2, 3, 1, "", "cdf_args"], [2, 3, 1, "", "contour_args"], [2, 3, 1, "", "data"], [2, 3, 1, "", "density_args"], [2, 3, 1, "", "export_args"], [2, 2, 1, "", "export_plot_to_file"], [2, 3, 1, "", "fig"], [2, 3, 1, "", "gen_plot_args"], [2, 2, 1, "", "generic_plot"], [2, 3, 1, "", "hist_args"], [2, 2, 1, "", "label_axes"], [2, 3, 1, "", "norm_cdf_args"], [2, 2, 1, "", "plot_box_plots"], [2, 2, 1, "", "plot_cdf"], [2, 2, 1, "", "plot_contour"], [2, 2, 1, "", "plot_density"], [2, 2, 1, "", "plot_histogram"], [2, 2, 1, "", "plot_normal_cdf"], [2, 2, 1, "", "plot_probability_plot"], [2, 2, 1, "", "plot_timeseries"], [2, 3, 1, "", "prob_plot_args"], [2, 2, 1, "", "set_defaults"], [2, 3, 1, "", "timeseries_args"], [2, 2, 1, "", "update_defaults"], [2, 3, 1, "", "util"]], "chartly.chartly.PlotUtilities": [[2, 2, 1, "", "standardize_dataset"]], "tests": [[6, 0, 0, "-", "test_chartly"]], "tests.test_chartly": [[6, 1, 1, "", "TestPlotting"]], "tests.test_chartly.TestPlotting": [[6, 2, 1, "", "setUp"], [6, 2, 1, "", "test_contour_data_length"], [6, 2, 1, "", "test_default"], [6, 2, 1, "", "test_gen_plot_data_length"], [6, 2, 1, "", "test_gen_plot_data_type"], [6, 2, 1, "", "test_multi_clear_axis"], [6, 2, 1, "", "test_mutli_subplot_count"], [6, 2, 1, "", "test_standardize_data"]]}, "objnames": {"0": ["py", "module", "Python module"], "1": ["py", "class", "Python class"], "2": ["py", "method", "Python method"], "3": ["py", "attribute", "Python attribute"], "4": ["py", "data", "Python data"]}, "objtypes": {"0": "py:module", "1": "py:class", "2": "py:method", "3": "py:attribute", "4": "py:data"}, "terms": {"": 2, "0": [0, 1, 2, 7], "1": [0, 1, 2, 4, 6, 7], "10": [1, 2], "100": [1, 7], "1000": [0, 1], "11": 7, "14": 7, "1d": [1, 6], "2": [0, 1, 2, 6], "20": [1, 2], "26": 7, "2d": [1, 2, 6], "3": [1, 2, 6, 7], "30": 1, "4": [2, 7], "5": 2, "50": 0, "500": 1, "8": 0, "9": [1, 2, 7], "A": [0, 1, 2, 6], "The": [0, 1, 2, 7], "There": [1, 2, 7], "To": 7, "_": 2, "__call__": 2, "across": 0, "add": [], "addit": 1, "against": 2, "all": [0, 2], "allow": [0, 1, 7], "also": 1, "an": [2, 6, 7], "ang": [2, 6], "ani": 1, "api": 7, "apopo": [2, 6], "approxim": 1, "ar": [1, 2, 6, 7], "arg": [0, 1, 2, 7], "argument": [0, 1], "arrai": [1, 2], "assist": [2, 6], "author": [2, 6], "auto": 4, "autoapi": 4, "auxillari": 2, "ax": [0, 1, 2, 7], "axes_label": [0, 1, 2], "axi": [0, 1, 2, 6, 7], "azenda": [2, 6], "base": [2, 6], "bin": [1, 2], "black": 2, "bool": 0, "both": [1, 2, 6], "box": 2, "boxplot": [1, 2], "boxplot_arg": [1, 2], "build": 2, "c": [2, 6], "calcul": 2, "call": 0, "can": [0, 1, 6, 7], "cdf": [1, 2], "cdf_arg": [1, 2], "chartli": 4, "cheikh": [2, 6], "class": [0, 1], "clear": 6, "clear_axi": 2, "cmbengu": [2, 6], "co": 1, "code": 7, "color": [1, 2], "color_list": 2, "column": 2, "com": [2, 6], "command": 7, "compani": [2, 6], "compar": 1, "consult": [2, 6], "contain": [1, 2, 4], "content": [], "contour": [2, 6], "contour_arg": [1, 2], "correctli": 6, "creat": [0, 1, 2, 4, 6, 7], "current": [1, 2], "current_subplot": 2, "custom": [1, 2], "data": [0, 1, 2, 6, 7], "dataset": [1, 2], "dataset_on": 1, "dataset_thre": 1, "dataset_two": 1, "decis": [2, 6], "default": [2, 6], "defin": 0, "densiti": [0, 1, 2], "density_arg": [1, 2], "detail": [], "determin": 1, "develop": [2, 6], "dict": 2, "dictionari": [0, 1, 2], "differ": 1, "displai": 2, "distribut": 2, "divid": 2, "document": 4, "e": [2, 6], "each": 2, "ec": [2, 6], "eci": [2, 6], "elizabeth": [2, 6], "ep": 2, "error": 6, "estim": 2, "etc": 1, "exampl": [2, 7], "exponenti": 1, "export": 2, "export_arg": 2, "export_plot_to_fil": 2, "fals": [0, 2], "field": 2, "fig": [1, 2, 7], "figur": [0, 1, 2, 7], "file": 2, "fill": [1, 2], "finish": 0, "firebrick": 1, "fname": 2, "follow": [1, 2, 7], "font": 2, "format": 2, "frequenc": 1, "from": 2, "fsize": 2, "function": 2, "gamma": 1, "gen_plot": 0, "gen_plot_arg": [1, 2], "gener": [2, 4, 6], "generic_plot": [1, 2, 7], "graph": 2, "guid": 7, "handl": 6, "here": 7, "hist_arg": [1, 2], "histogram": [0, 2], "histogram_arg": 1, "hold": 2, "how": 7, "i": [0, 1, 2, 6, 7], "import": [0, 1, 7], "inc": [2, 6], "index": 7, "indi": [2, 6], "inlin": 2, "input": 2, "int": 2, "intern": [2, 6], "intl": [2, 6], "its": 2, "just": 1, "k": 2, "kei": [1, 2], "kernel": 2, "label": [0, 1, 2], "label_ax": 2, "legend": 2, "length": 6, "lightpink": 2, "line": [2, 7], "linear": 2, "linelabel": 2, "linestyl": [1, 2], "linspac": 1, "list": [0, 2, 6], "loc": [0, 1], "lucia": [2, 6], "m": [2, 6], "main": [0, 2, 7], "mari": [2, 6], "master": 2, "matplotlib": [1, 7], "mbengu": [2, 6], "mean": 2, "mediumpurpl": 1, "mediumvioletr": 2, "meshgrid": 1, "method": [0, 1, 2, 7], "methodnam": 6, "minim": 7, "modul": 7, "multi": 0, "multipl": [0, 1, 2, 6, 7], "multiplot": [2, 6, 7], "must": 2, "name": 2, "navi": 2, "ndarrai": 2, "need": 7, "new": [0, 2], "new_subplot": [0, 2], "new_valu": 2, "none": 2, "norm_cdf": 0, "norm_cdf_arg": [1, 2], "normal": [0, 2], "notch": 1, "note": 2, "np": [0, 1, 7], "num": 2, "num_bin": [1, 2], "number": 2, "numpi": [2, 7], "o": [2, 6], "object": [0, 1, 6, 7], "onli": 1, "onto": 2, "option": 2, "orang": 2, "organ": [2, 6], "other": 2, "oumar": [2, 6], "outlier": 2, "overlai": [2, 7], "overlay_arg": 2, "overlay_payload": 0, "packag": 7, "page": [4, 7], "paramet": 2, "pass": [0, 1], "payload": 0, "pip": 7, "pleas": 2, "plot": [2, 6, 7], "plot_box_plot": 2, "plot_cdf": 2, "plot_contour": [1, 2], "plot_contour_plot": [], "plot_dens": 2, "plot_histogram": [1, 2], "plot_normal_cdf": [1, 2], "plot_probability_plot": [1, 2], "plot_timeseri": 2, "plottingutil": 2, "plotutil": 2, "plt": [1, 2, 7], "plum": 2, "popo": [2, 6], "privat": [2, 6], "prob_plot": 0, "prob_plot_arg": [1, 2], "probabl": 2, "pyplot": [1, 7], "ran": 2, "randn": [1, 7], "random": [0, 1, 7], "rang": 2, "red": [1, 2], "refer": 7, "render": 0, "requir": [0, 1, 2], "research": [2, 6], "reset": 2, "restructuredtext": [], "result": 2, "return": 2, "row": 2, "run": 7, "runtest": 6, "salmon": 1, "same": 2, "scale": [0, 1, 2], "scientist": [2, 6], "scipi": 7, "seaborn": 7, "search": 7, "sector": [2, 6], "see": [], "seri": 1, "set": [0, 1, 2, 6], "set_default": 2, "setup": 6, "share": [0, 2], "share_ax": [0, 2], "show": [1, 2, 7], "show_legend": 2, "showflier": 2, "simpl": [1, 7], "sin": 1, "singl": [0, 1, 7], "size": [0, 1, 2], "skyblu": 2, "slateblu": 2, "solid": 2, "some": 0, "special": [2, 6], "specifi": [1, 7], "sphinx": 4, "st": [2, 6], "standard": [2, 6], "standardize_dataset": 2, "std": 2, "str": [0, 2], "style": 2, "subplot": [1, 2, 6, 7], "subplot_count": 2, "subtract": 2, "super_titl": [0, 2], "super_xlabel": [0, 2], "super_ylabel": [0, 2], "support": [2, 6], "syntax": [], "system": [2, 6], "test": [2, 4, 7], "test_chartli": [4, 5], "test_contour_data_length": 6, "test_default": 6, "test_gen_plot_data_length": 6, "test_gen_plot_data_typ": 6, "test_multi_clear_axi": 6, "test_mutli_subplot_count": 6, "test_plot": [], "test_standardize_data": 6, "testcas": 6, "testplot": 6, "thei": 0, "them": 0, "thi": [2, 4, 7], "throw": 6, "tile": 2, "time": [1, 7], "timeseri": 2, "timeseries_arg": [1, 2], "titl": [0, 1, 2], "tool": 7, "tracker": 2, "true": 2, "tupl": 2, "two": 7, "type": [1, 2], "unequ": 6, "unittest": 6, "up": [0, 6], "updat": 2, "update_default": 2, "us": [1, 2, 6, 7], "usag": 2, "user": [0, 1], "util": 2, "valu": [1, 2], "variou": 2, "west": [2, 6], "when": 0, "whether": [0, 2], "x": [0, 1, 2], "xlabel": [0, 1, 2], "y": [0, 1, 2], "ylabel": [0, 1, 2], "you": 7, "your": [], "z": 1}, "titles": ["Chartly.Multiplots Usage Guide", "chartly.Plot Usage Guide", "chartly.chartly", "chartly", "API Reference", "tests", "tests.test_chartly", "Chartly"], "titleterms": {"api": 4, "attribut": 2, "chartli": [0, 1, 2, 3, 7], "class": [2, 6], "content": [2, 6], "contour": 1, "cumul": 1, "distribut": 1, "document": [], "exampl": [0, 1], "function": 1, "gener": 1, "guid": [0, 1], "histogram": 1, "indic": 7, "instal": 7, "modul": [2, 6], "multiplot": 0, "normal": 1, "overlai": 0, "overview": 7, "plot": [0, 1], "probabl": 1, "refer": 4, "requir": 7, "submodul": [3, 5], "subplot": 0, "tabl": 7, "test": [5, 6], "test_chartli": 6, "test_plot": [], "usag": [0, 1, 7]}})
